requirements.txt
# Core dependencies
langchain>=0.1.0
openai>=1.0.0
pinecone-client>=2.0.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# ML/AI
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
accelerate>=0.24.0
peft>=0.6.0

# Data collection
alpha-vantage>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Caching
redis>=5.0.0

# Sentiment analysis
finbert-embedding>=0.1.0

